tmux: support running commands for not-found shortcuts
If the "Any" shortcut key is specified,
and a user types any part of a shortcut that does not exist,
the tmux command specified as the value will be executed.
This will not happen, however, if the first character of a shortcut is typed incorrectly.
This avoids altering the root table more than absolutely necissary.

Example:
f3 prz
will execute the Any action when the user types z
f2 prod
will abort silently
f3 prod
will send the shortcut for prod