appalachian-power-history-from-fiddler: extract quarter hour power readings
This pulls readings from an extracted fiddler saz file.
Useful if you want to see why/when your bills have skyrocketed.
There is another script on GH for Opower data retrieval,
but at the time of this commit it does not use the graphql API in use by Appalachian Power.